+++ packages/sdk-coverage-tests/src/common-util.js
@@ -1,7 +1,7 @@
 const vm = require('vm')
 const fs = require('fs')
-const fetch = require('node-fetch')
 const chalk = require('chalk')
+const fetchSync = require('sync-fetch')
 
 function findDifferencesBetweenCollections(hostCollection = [], guestCollection = []) {
   const _hostCollection = Array.isArray(hostCollection)
@@ -59,23 +59,38 @@ function toPascalCase(string) {
     .join('')
 }
 
-async function requirePath(path, context) {
-  const source = isUrl(path)
-    ? await fetch(path).then(response => response.text())
-    : fs.readFileSync(path).toString()
+function loadFile(path) {
+  return isUrl(path) ? fetchSync(path).text() : fs.readFileSync(path).toString()
+}
 
+function runCode(code, context) {
   try {
-    return vm.runInContext(source, vm.createContext({...context, process}))
+    return vm.runInContext(code, vm.createContext({...context, console, process}))
   } catch (err) {
     if (err.constructor.name !== 'ReferenceError') throw err
     const stack = err.stack.split('\n')
     const [columnNumber, lineNumber] = stack[5].split(':').reverse()
-    console.log(chalk.yellow(`Error during requiring ${path}:${columnNumber}:${lineNumber}`), '\n')
+    console.log(chalk.yellow(`Error during running ${columnNumber}:${lineNumber}`), '\n')
     const [line, caret] = stack.slice(1, 3)
     console.log(chalk.cyan(line))
     console.log(chalk.yellow(caret))
     throw new ReferenceError(err.message)
   }
+}
+
+async function requireUrl(url, cache = {}) {
+  const code = loadFile(url)
+  const module = {exports: {}}
+  cache[url] = module
+  runCode(code, {
+    require(path) {
+      if (!/^[./]/.test(path)) return require(path)
+      const requiringUrl = new URL(path, url).href
+      return cache[requiringUrl] ? cache[requiringUrl].exports : requireUrl(requiringUrl, cache)
+    },
+    module,
+  })
+  return module.exports
 }
 
 module.exports = {
@@ -88,5 +103,7 @@ module.exports = {
   isEmptyObject,
   capitalize,
   toPascalCase,
-  requirePath,
+  loadFile,
+  runCode,
+  requireUrl,
 }